Add additional columns
