+++ package.json
@@ -83,7 +83,7 @@
     "webpack-merge": "^5.7.3"
   },
   "dependencies": {
-    "@centreon/ui": "github:centreon/centreon-ui#make-listing-columns-customizable",
+    "@centreon/ui": "github:centreon/centreon-ui",
     "@centreon/ui-context": "github:centreon/ui-context",
     "@date-io/dayjs": "1.3.13",
     "@material-ui/core": "4.11.3",
