+++ www/front_src/src/Resources/Listing/columns/IconColumn.tsx
@@ -1,0 +1,23 @@
+import * as React from 'react';
+
+import { makeStyles } from '@material-ui/styles';
+
+interface Props {
+  children: React.ReactNode;
+}
+
+const useStyles = makeStyles(() => ({
+  column: {
+    display: 'flex',
+    justifyContent: 'center',
+    width: '100%',
+  },
+}));
+
+const IconColumn = ({ children }: Props): JSX.Element => {
+  const classes = useStyles();
+
+  return <div className={classes.column}>{children}</div>;
+};
+
+export default IconColumn;
