+++ www/front_src/src/Resources/Listing/columns/Severity.tsx
@@ -1,0 +1,28 @@
+import * as React from 'react';
+
+import { makeStyles } from '@material-ui/core';
+
+import { ComponentColumnProps, SeverityCode, StatusChip } from '@centreon/ui';
+
+import { useColumnStyles } from '.';
+
+const SeverityColumn = ({ row }: ComponentColumnProps): JSX.Element | null => {
+  const classes = useColumnStyles();
+
+  if (!row.severity_level) {
+    return null;
+  }
+
+  return (
+    <StatusChip
+      label={row.severity_level?.toString()}
+      severityCode={SeverityCode.None}
+      classes={{
+        root: classes.extraSmallChipContainer,
+        label: classes.smallChipLabel,
+      }}
+    />
+  );
+};
+
+export default SeverityColumn;
