+++ www/front_src/src/Resources/Listing/columns/Url/index.tsx
@@ -6,6 +6,7 @@ import { useTranslation } from 'react-i18next';
 import { IconButton } from '@centreon/ui';
 
 import { labelUrl } from '../../../translatedLabels';
+import IconColumn from '../IconColumn';
 
 interface Props {
   endpoint?: string;
@@ -21,22 +22,24 @@ const UrlColumn = ({ endpoint, title, icon }: Props): JSX.Element | null => {
   }
 
   return (
-    <a
-      href={endpoint}
-      onClick={(e): void => {
-        e.stopPropagation();
-      }}
-    >
-      <IconButton
-        title={t(title || labelUrl)}
-        ariaLabel={title}
-        onClick={(): null => {
-          return null;
+    <IconColumn>
+      <a
+        href={endpoint}
+        onClick={(e): void => {
+          e.stopPropagation();
         }}
       >
-        {icon}
-      </IconButton>
-    </a>
+        <IconButton
+          title={t(title || labelUrl)}
+          ariaLabel={title}
+          onClick={(): null => {
+            return null;
+          }}
+        >
+          {icon}
+        </IconButton>
+      </a>
+    </IconColumn>
   );
 };
