+++ www/front_src/src/Resources/Listing/columns/index.tsx
@@ -1,22 +1,9 @@
-import React from 'react';
+import { pipe, split, head, propOr, T } from 'ramda';
 
-import { pipe, split, head, propOr, T, pathEq } from 'ramda';
-import { useTranslation } from 'react-i18next';
+import { makeStyles } from '@material-ui/core';
 
-import { Typography, makeStyles } from '@material-ui/core';
-import IconAcknowledge from '@material-ui/icons/Person';
-import IconCheck from '@material-ui/icons/Sync';
+import { ColumnType, Column } from '@centreon/ui';
 
-import {
-  ColumnType,
-  StatusChip,
-  SeverityCode,
-  IconButton,
-  Column,
-  ComponentColumnProps,
-} from '@centreon/ui';
-
-import IconDowntime from '../../icons/Downtime';
 import {
   labelResource,
   labelStatus,
@@ -25,10 +12,6 @@ import {
   labelInformation,
   labelState,
   labelLastCheck,
-  labelAcknowledge,
-  labelSetDowntimeOn,
-  labelCheck,
-  labelSetDowntime,
   labelParent,
   labelNotes,
   labelAction,
@@ -36,14 +19,21 @@ import {
   labelAlias,
   labelFqdn,
   labelMonitoringServer,
+  labelNotification,
+  labelCheck,
 } from '../../translatedLabels';
-import useAclQuery from '../../Actions/Resource/aclQuery';
 import truncate from '../../truncate';
 
 import StateColumn from './State';
 import GraphColumn from './Graph';
 import NotesUrlColumn from './Url/Notes';
 import ActionUrlColumn from './Url/Action';
+import StatusColumn from './Status';
+import SeverityColumn from './Severity';
+import ResourceColumn from './Resource';
+import ParentResourceColumn from './Parent';
+import NotificationColumn from './Notification';
+import ChecksColumn from './Checks';
 
 const useStyles = makeStyles((theme) => ({
   resourceDetailsCell: {
@@ -67,11 +57,6 @@ const useStyles = makeStyles((theme) => ({
   smallChipLabel: {
     padding: theme.spacing(0.5),
   },
-  statusColumn: {
-    display: 'flex',
-    alignItems: 'center',
-    width: '100%',
-  },
   actions: {
     display: 'flex',
     flexWrap: 'nowrap',
@@ -81,147 +66,7 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const SeverityColumn = ({ row }: ComponentColumnProps): JSX.Element | null => {
-  const classes = useStyles();
-
-  if (!row.severity_level) {
-    return null;
-  }
-
-  return (
-    <StatusChip
-      label={row.severity_level?.toString()}
-      severityCode={SeverityCode.None}
-      classes={{
-        root: classes.extraSmallChipContainer,
-        label: classes.smallChipLabel,
-      }}
-    />
-  );
-};
-
-type StatusColumnProps = {
-  actions;
-} & Pick<ComponentColumnProps, 'row'>;
-
-const StatusColumnOnHover = ({
-  actions,
-  row,
-}: StatusColumnProps): JSX.Element => {
-  const classes = useStyles();
-  const { t } = useTranslation();
-
-  const { canAcknowledge, canDowntime, canCheck } = useAclQuery();
-
-  const isResourceOk = pathEq(
-    ['status', 'severity_code'],
-    SeverityCode.Ok,
-    row,
-  );
-
-  const disableAcknowledge = !canAcknowledge([row]) || isResourceOk;
-  const disableDowntime = !canDowntime([row]);
-  const disableCheck = !canCheck([row]);
-
-  return (
-    <div className={classes.actions}>
-      <IconButton
-        title={t(labelAcknowledge)}
-        disabled={disableAcknowledge}
-        color="primary"
-        onClick={(): void => actions.onAcknowledge(row)}
-        ariaLabel={`${t(labelAcknowledge)} ${row.name}`}
-      >
-        <IconAcknowledge fontSize="small" />
-      </IconButton>
-      <IconButton
-        title={t(labelSetDowntime)}
-        disabled={disableDowntime}
-        onClick={(): void => actions.onDowntime(row)}
-        ariaLabel={`${t(labelSetDowntimeOn)} ${row.name}`}
-      >
-        <IconDowntime fontSize="small" />
-      </IconButton>
-      <IconButton
-        title={t(labelCheck)}
-        disabled={disableCheck}
-        onClick={(): void => actions.onCheck(row)}
-        ariaLabel={`${t(labelCheck)} ${row.name}`}
-      >
-        <IconCheck fontSize="small" />
-      </IconButton>
-    </div>
-  );
-};
-
-const StatusColumn = ({ actions, t }) => {
-  const Status = ({ row, isHovered }: ComponentColumnProps): JSX.Element => {
-    const classes = useStyles();
-
-    const statusName = row.status.name;
-
-    return (
-      <div className={classes.statusColumn}>
-        {isHovered ? (
-          <StatusColumnOnHover actions={actions} row={row} />
-        ) : (
-          <StatusChip
-            style={{ height: 20, width: '100%' }}
-            label={t(statusName)}
-            severityCode={row.status.severity_code}
-          />
-        )}
-      </div>
-    );
-  };
-
-  return Status;
-};
-
-const ResourceColumn = ({ row }: ComponentColumnProps): JSX.Element => {
-  const classes = useStyles();
-
-  return (
-    <div className={classes.resourceDetailsCell}>
-      {row.icon ? (
-        <img src={row.icon.url} alt={row.icon.name} width={16} height={16} />
-      ) : (
-        <StatusChip
-          label={row.short_type}
-          severityCode={SeverityCode.None}
-          classes={{
-            root: classes.extraSmallChipContainer,
-            label: classes.smallChipLabel,
-          }}
-        />
-      )}
-      <div className={classes.resourceNameItem}>
-        <Typography variant="body2">{row.name}</Typography>
-      </div>
-    </div>
-  );
-};
-
-const ParentResourceColumn = ({
-  row,
-}: ComponentColumnProps): JSX.Element | null => {
-  const classes = useStyles();
-
-  if (!row.parent) {
-    return null;
-  }
-
-  return (
-    <div className={classes.resourceDetailsCell}>
-      <StatusChip severityCode={row.parent?.status?.severity_code || 0} />
-      <div className={classes.resourceNameItem}>
-        <Typography variant="body2">{row.parent.name}</Typography>
-      </div>
-    </div>
-  );
-};
-
-interface ColumnsProps {
+export interface ColumnProps {
   actions;
   t: (value: string) => string;
 }
@@ -241,7 +86,7 @@ export const defaultSelectedColumnIds = [
   'state',
 ];
 
-export const getColumns = ({ actions, t }: ColumnsProps): Array<Column> => [
+export const getColumns = ({ actions, t }: ColumnProps): Array<Column> => [
   {
     id: 'severity',
     label: 'S',
@@ -360,4 +205,20 @@ export const getColumns = ({ actions, t }: ColumnsProps): Array<Column> => [
     getFormattedString: ({ monitoring_server_name }): string =>
       monitoring_server_name,
   },
+  {
+    id: 'notification',
+    label: t(labelNotification),
+    type: ColumnType.component,
+    getRenderComponentOnRowUpdateCondition: T,
+    Component: NotificationColumn,
+  },
+  {
+    id: 'checks',
+    label: t(labelCheck),
+    type: ColumnType.component,
+    getRenderComponentOnRowUpdateCondition: T,
+    Component: ChecksColumn,
+  },
 ];
+
+export { useStyles as useColumnStyles };
