+++ www/front_src/src/Resources/translatedLabels.ts
@@ -200,3 +200,6 @@ export const labelEndDateGreaterThanStartDate =
   'The end date must be greater than the start date';
 export const labelAction = 'Action';
 export const labelNotes = 'Notes';
+export const labelChecksDisabled = 'Checks disabled';
+export const labelNotificationDisabled = 'Notifications disabled';
+export const labelNotificationEnabled = 'Notifications enabled';
